requirements.txt
# Environment variables management
python-dotenv == 1.0.1

# Certificates
certifi == 2024.12.14 

# Character encoding
charset-normalizer == 3.4.1 

# HTTP libraries
idna == 3.10 
requests == 2.32.3 
urllib3 == 2.3.0
requests-oauthlib == 2.0.0

# Logging
loguru == 0.7.3

# Asynchronous networking
aiohappyeyeballs == 2.4.4
aiohttp == 3.11.11
aiosignal == 1.3.2 
async-timeout == 5.0.1 

# Utility libraries
attrs == 24.3.0 
typing-extensions == 4.12.2 

# Discord API
discord == 2.3.2 
discord.py == 2.4.0 

# Data structures
frozenlist == 1.5.0 
multidict == 6.1.0 
yarl == 1.18.3

# Caching
propcache == 0.2.1 
cachetools == 5.5.1

# Google authentication and API
google-auth == 2.38.0
google-auth-oauthlib == 1.2.1
gspread == 6.1.4
oauthlib == 3.2.2
pyasn1 == 0.6.1
pyasn1-modules == 0.4.1
rsa == 4.9